get dataset when processing single promise only

+++ src/config.js
@@ -28,7 +28,7 @@ const config = {
   title: "PromiseTracker",
   URL: "https://promisetracker.dev.codeforafrica.org",
   DATE_OPTIONS: {
-    day: "short",
+    day: "numeric",
     month: "long",
     year: "numeric",
   },
